// src/app/services/sandwich-client/model/ingredienteDTo.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface IngredienteDTo { 
    id?: number;
    nombre?: string;
    esVegetariano?: boolean;
}