// src/app/services/sandwich-client/model/sandwichRequestDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SandwichRequestDTO { 
    nombre?: string;
    descripcion?: string;
    tipoId?: number;
    paisId?: number;
    ingredientesIds?: Array<number>;
    vegetariano?: boolean;
    caloriasAproximadas?: number;
    precio?: number;
    disponible?: boolean;
    imagenUrl?: string;
}