// src/app/services/sandwich-client/model/tipoDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TipoDto { 
    id?: number;
    nombre?: string;
    descripcion?: string;
    fechaCreacion?: string;
    fechaModificacion?: string;
}